// payload-types.ts
export interface User {
  id: number
  name: string
  updatedAt: string
  createdAt: string
  email: string
  resetPasswordToken?: string | null
  resetPasswordExpiration?: string | null
  salt?: string | null
  hash?: string | null
  loginAttempts?: number | null
  lockUntil?: string | null
  password?: string | null
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number
  alt?: string | null
  caption?: {
    root: {
      type: string
      children: {
        type: string
        version: number
        [k: string]: unknown
      }[]
      direction: ("ltr" | "rtl") | null
      format: "left" | "start" | "center" | "right" | "end" | "justify" | ""
      indent: number
      version: number
    }
    [k: string]: unknown
  } | null
  updatedAt: string
  createdAt: string
  url?: string | null
  thumbnailURL?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
  focalX?: number | null
  focalY?: number | null
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "videos".
 */
export interface Video {
  id: number
  caption?: {
    root: {
      type: string
      children: {
        type: string
        version: number
        [k: string]: unknown
      }[]
      direction: ("ltr" | "rtl") | null
      format: "left" | "start" | "center" | "right" | "end" | "justify" | ""
      indent: number
      version: number
    }
    [k: string]: unknown
  } | null
  updatedAt: string
  createdAt: string
  url?: string | null
  thumbnailURL?: string | null
  filename?: string | null
  mimeType?: string | null
  filesize?: number | null
  width?: number | null
  height?: number | null
  focalX?: number | null
  focalY?: number | null
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "articles".
 */
export interface Article {
  id: number
  title: string
  /**
   * Featured image of the article
   */
  illustration?: (number | null) | Media
  content: {
    root: {
      type: string
      children: {
        type: string
        version: number
        [k: string]: unknown
      }[]
      direction: ("ltr" | "rtl") | null
      format: "left" | "start" | "center" | "right" | "end" | "justify" | ""
      indent: number
      version: number
    }
    [k: string]: unknown
  }
  meta?: {
    title?: string | null
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media
    description?: string | null
  }
  slug: string
  slugLock?: boolean | null
  publishedAt?: string | null
  author?: (number | null) | User
  updatedAt: string
  createdAt: string
  _status?: ("draft" | "published") | null
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "projects".
 */
export interface Project {
  id: number
  title: string
  /**
   * Video preview of the project
   */
  video?: (number | null) | Video
  /**
   * Source of the video
   */
  videoSrc?: string | null
  Content: {
    content: {
      root: {
        type: string
        children: {
          type: string
          version: number
          [k: string]: unknown
        }[]
        direction: ("ltr" | "rtl") | null
        format: "left" | "start" | "center" | "right" | "end" | "justify" | ""
        indent: number
        version: number
      }
      [k: string]: unknown
    }
  }
  meta?: {
    title?: string | null
    /**
     * Maximum upload file size: 12MB. Recommended file size for images is <500KB.
     */
    image?: (number | null) | Media
    description?: string | null
  }
  url?: string | null
  githubUrl?: string | null
  slug: string
  slugLock?: boolean | null
  updatedAt: string
  createdAt: string
}